Start internal ElementHandlerUtil class
