+++ kotlinpoet-metadata-specs/src/main/kotlin/com/squareup/kotlinpoet/metadata/specs/ConstructorData.kt
@@ -3,6 +3,7 @@ package com.squareup.kotlinpoet.metadata.specs
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.metadata.KotlinPoetMetadataPreview
+import com.squareup.kotlinpoet.metadata.specs.internal.ElementHandlerUtil
 import com.squareup.kotlinpoet.metadata.specs.internal.JVM_SYNTHETIC_SPEC
 
 /**
@@ -31,7 +32,7 @@ data class ConstructorData(
   val allAnnotations: Collection<AnnotationSpec> = ElementHandler.createAnnotations {
     addAll(annotations)
     if (isSynthetic) {
-      add(JVM_SYNTHETIC_SPEC)
+      add(ElementHandlerUtil.JVM_SYNTHETIC_SPEC)
     }
     addAll(jvmModifiers.map { it.annotationSpec() })
     exceptions.takeIf { it.isNotEmpty() }
