+++ kotlinpoet-metadata-specs/src/main/kotlin/com/squareup/kotlinpoet/metadata/specs/internal/KmTypes.kt
@@ -15,7 +15,6 @@
  */
 package com.squareup.kotlinpoet.metadata.specs.internal
 
-import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.LambdaTypeName
@@ -25,7 +24,6 @@ import com.squareup.kotlinpoet.STAR
 import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.TypeVariableName
 import com.squareup.kotlinpoet.WildcardTypeName
-import com.squareup.kotlinpoet.asClassName
 import com.squareup.kotlinpoet.metadata.ImmutableKmClass
 import com.squareup.kotlinpoet.metadata.ImmutableKmConstructor
 import com.squareup.kotlinpoet.metadata.ImmutableKmFlexibleTypeUpperBound
@@ -172,13 +170,3 @@ private fun ImmutableKmFlexibleTypeUpperBound.toTypeName(
   return WildcardTypeName.producerOf(type.toTypeName(typeParamResolver))
 }
 
-internal val JVM_FIELD = JvmField::class.asClassName()
-internal val JVM_FIELD_SPEC = AnnotationSpec.builder(JVM_FIELD).build()
-internal val JVM_SYNTHETIC_SPEC = AnnotationSpec.builder(JvmSynthetic::class).build()
-internal val JVM_TRANSIENT = Transient::class.asClassName()
-internal val JVM_VOLATILE = Volatile::class.asClassName()
-internal val IMPLICIT_FIELD_ANNOTATIONS = setOf(
-    JVM_FIELD,
-    JVM_TRANSIENT,
-    JVM_VOLATILE
-)
